adding AuthController For Authentication Process
